pre-upload: Add comment to --ignore=GERRIT_CHANGE_ID option
This check for Change-Id: is present in checkpatch.pl since kernel 3.15,
introduced in commit 7ebd05ef1646e8cbef54e38343722741a4744626 upstream.

For upstream kernel, Change-Id: is noise, but as we do use Gerrit to
track commits in all our repos, it is actually required for us.
Disable the checkpatch.pl check for it to suppress the warning.

BUG=chromium:533701
TEST=Presubmit passes.

Change-Id: Iec8d9dd4ec0b3c95eb9559d95d01996757840055
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/304866
Reviewed-by: Mike Frysinger <vapier@chromium.org>