pre-upload: allow some more tags in cherry picked patches
It turns out a couple of common tags were missed when the list of the
tags allowed in the last paragraph of the commit message was defined.

'Commit-Queue:' added by gerrit when the original patch is checked in
in some cases, and '(cherry picked from commit...' which can be
showing up more than once in case a patch was cherry picked more than
once.

Let's add these to the list of allowed tags to avoid false positives
when cherry picking into branches.

BUG=none
TEST=verified that proper error message is generated for different
     inputs, and that valid tags do not trigger pre-submit errors any
     more.

Change-Id: I4ba4259982abe61fa66cbc064ddee1418fd0b051
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/666490
Reviewed-by: Guenter Roeck <groeck@chromium.org>